requirements.txt
psycopg2-binary
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
matplotlib>=3.7.0
plotly
requests>=2.28.0
schedule>=1.2.0
python-dotenv>=1.0.0
reportlab
schedule
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0








